[BREAKS CI] Add Gemfile generated by test-deploy to repo
This will be used late to dynamically generate a dummy extension.

We're choosing to commit this Gemfile and its lock file to the repo so
that our gem environment would be deterministic across builds.

This breaks the CI because test-deploy.yml currently creates a new
Gemfile when creating a dummy extension.