Allow extensions to invalidate bundle cache
Until now, there were no way for an extension to invalidate its bundle
cache. Instead, it was needed to bump the cache version upstream in the
orb's code. That also meant that all of the extensions cache was
invalidated at once.

We're now giving control to individual extensions by adding a
`CACHE_VERSION` environment variable to the cache keys [1]. When no
present in the extension's environment, it'll take a value equal to `<no
value>`, and
it can be added and bumped to invalidate it.

[1] - https://support.circleci.com/hc/en-us/articles/115015426888-Clear-Project-Dependency-Cache